requirements.txt
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
emoji==2.12.1
exceptiongroup==1.2.2
flake8==7.1.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
macholib==1.16.3
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pygame==2.6.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PySocks==1.7.1
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
